Revert "Refactor Dockerfile, workflows, and filters; add .dockerignore (#3558)"
This reverts commit e4bb4507a8654cf754acfa1c112c40c657f338b5.
