--- .dockerignore
@@ -1,8 +1,0 @@
-.git
-.idea
-target
-**/*.iml
-.mvn/wrapper
-node_modules
-*.log
-.DS_Store
