+++ .github/workflows/start-bot.yml
@@ -10,6 +10,8 @@ jobs:
     - name: executing remote ssh commands using password
       uses: appleboy/ssh-action@master
       env:
+        AWS_KEY: ${{ secrets.IMG_AWS_KEY }}
+        AWS_SECRET: ${{ secrets.IMG_AWS_SECRET }}
         DISCORD_BOT_TOKEN: ${{ secrets.DISCORD_BOT_TOKEN }}
         DISCORD_BOT_USERID: ${{ secrets.DISCORD_BOT_USERID }}
         REPO_DISPATCH_TOKEN: ${{ secrets.REPOSITORY_DISPATCH_TOKEN }}
@@ -20,14 +22,14 @@ jobs:
         username: ${{ secrets.HOSTINGER_SSH_USER }}
         password: ${{ secrets.HOSTINGER_SSH_PASSWORD }}
         port: ${{ secrets.HOSTINGER_SSH_PORT }}
-        envs: DISCORD_BOT_TOKEN, DISCORD_BOT_USERID, REPO_DISPATCH_TOKEN, TI4_ULTIMATE_STATISTICS_API_KEY, MANAGEMENT_ENDPOINTS_ACTUATOR_API_KEY
+        envs: AWS_KEY, AWS_SECRET, DISCORD_BOT_TOKEN, DISCORD_BOT_USERID, REPO_DISPATCH_TOKEN, TI4_ULTIMATE_STATISTICS_API_KEY, MANAGEMENT_ENDPOINTS_ACTUATOR_API_KEY
         script: |
           cd ${{ vars.HOST_TI4_REPO_DIR }}
           echo "Building docker image..."
           docker version
-          docker build -t tibot .
+          docker build --build-arg BOT_KEY=$DISCORD_BOT_KEY --build-arg USER=$DISCORD_USER --build-arg SERVER=$DISCORD_SERVER --build-arg SERVER2=$DISCORD_SERVER2 --build-arg AWS_SECRET=$AWS_SECRET --build-arg AWS_KEY=$AWS_KEY -t tibot .
           echo "Starting Container..."
-          docker run -v ${{ vars.HOST_TI4_SAVES_DIR }}:/opt/STORAGE -p 8081:8081 -d --restart unless-stopped -m 7g -e TI4_ULTIMATE_STATISTICS_API_KEY=$TI4_ULTIMATE_STATISTICS_API_KEY -e MANAGEMENT_ENDPOINTS_ACTUATOR_API_KEY=$MANAGEMENT_ENDPOINTS_ACTUATOR_API_KEY -e REPO_DISPATCH_TOKEN=$REPO_DISPATCH_TOKEN -e AWS_ACCESS_KEY_ID=$IMG_AWS_KEY -e AWS_SECRET_ACCESS_KEY=$IMG_AWS_SECRET tibot $DISCORD_BOT_TOKEN $DISCORD_BOT_USERID 943410040369479690 847560709730730064 1062139934745559160 1090910555327434774 1145823841227112598 1176104225932058694 1209956332380229672 1250131684393881610 1312882116597518416 1378702133297414164 1218341032857440259
+          docker run -v ${{ vars.HOST_TI4_SAVES_DIR }}:/opt/STORAGE -p 8081:8081 -d --restart unless-stopped -m 7g -e TI4_ULTIMATE_STATISTICS_API_KEY=$TI4_ULTIMATE_STATISTICS_API_KEY -e MANAGEMENT_ENDPOINTS_ACTUATOR_API_KEY=$MANAGEMENT_ENDPOINTS_ACTUATOR_API_KEY -e REPO_DISPATCH_TOKEN=$REPO_DISPATCH_TOKEN tibot $DISCORD_BOT_TOKEN $DISCORD_BOT_USERID 943410040369479690 847560709730730064 1062139934745559160 1090910555327434774 1145823841227112598 1176104225932058694 1209956332380229672 1250131684393881610 1312882116597518416 1378702133297414164 1218341032857440259
           echo "Cleaning up docker stuff..."
           docker rm $(docker ps --filter status=exited -q) || true
           docker rmi -f $(docker images --filter "dangling=true" -q) || true
